+++ weather-service/src/main/kotlin/weather/util/metrics/CallMetrics.kt
@@ -3,5 +3,5 @@ package weather.util.metrics
 class CallMetrics(
         val callType: String,
         val method: String,
-        val responseTime: Long,
+        val responseTime: Double,
         val statusCode: String)
